All floats to doubles (what was I thinking)

+++ src/ee/liiser/siim/galaxies/calculations/Calculator.java
@@ -1,6 +1,6 @@
 package ee.liiser.siim.galaxies.calculations;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 import ee.liiser.siim.galaxies.data.BaseObject;
 import ee.liiser.siim.galaxies.data.Core;
@@ -23,7 +23,7 @@ public class Calculator {
 	 * Timestep of the calculation.
 	 * Decrease this to improve accuracy, increase this to improve performance
 	 */
-	public static float dt = 0.001f;
+	public static double dt = 0.001;
 
 	/**
 	 * Enum holding the possible calculation methods
@@ -73,18 +73,18 @@ public class Calculator {
 	 */
 	private void updatePos(VelocityObject point) {
 		// Velocity Verlet'
-		Vector3f v05 = new Vector3f();
+		Vector3d v05 = new Vector3d();
 		v05.scale(dt / 2, acc(point));
 		v05.add(point.getVelocity());
 		// v(t+dt/2) = v + a*dt/2
 
-		Vector3f newPos2 = new Vector3f();
+		Vector3d newPos2 = new Vector3d();
 		newPos2.scale(dt, v05);
 		newPos2.add(point.getPosition());
 		point.setPosition(newPos2);
 		// x(t+dt) = x + v05*dt
 
-		Vector3f newVel = new Vector3f();
+		Vector3d newVel = new Vector3d();
 		newVel.scale(dt / 2, acc(point));
 		newVel.add(v05);
 		point.setVelocity(newVel);
@@ -98,10 +98,10 @@ public class Calculator {
 	 */
 	private void updatePos(BasicObject point) {
 		// Basic Verlet':
-		Vector3f newPos = new Vector3f();
+		Vector3d newPos = new Vector3d();
 		newPos.scale(2, point.getPosition());
 		newPos.sub(point.getOldPosition());
-		Vector3f a = acc(point);
+		Vector3d a = acc(point);
 		a.scale(dt * dt);
 		newPos.add(a);
 		// newPos = 2*pos - oldPos + acc(pos)*dt^2
@@ -116,17 +116,17 @@ public class Calculator {
 	 * @param point to find the acceleration for
 	 * @return the acceleration acting upon the point
 	 */
-	private Vector3f acc(BaseObject point) {
-		Vector3f a = new Vector3f();
+	private Vector3d acc(BaseObject point) {
+		Vector3d a = new Vector3d();
 		for (Core core : cores) {
 			if (core == point)
 				continue;
-			Vector3f relPos = new Vector3f();
+			Vector3d relPos = new Vector3d();
 			// r.12
 			relPos.sub(((Drawable) core).getPosition(), point.getPosition());
 
 			// m * r12 / len(r12)^3
-			relPos.scale((float) (core.getMass() / Math.pow(relPos.length(), 3)));
+			relPos.scale((double) (core.getMass() / Math.pow(relPos.length(), 3)));
 
 			// a +=
 			a.add(relPos);
