+++ src/ee/liiser/siim/galaxies/data/Core.java
@@ -10,16 +10,16 @@ public interface Core extends Drawable{
 	/**
 	 * Visual size of the galaxy cores.
 	 */
-	public static final float CORE_SIZE = 0.05f;
+	public static final double CORE_SIZE = 0.05f;
 	/**
 	 * The default mass of galaxies. Change to modify the strength of the gravity
 	 */
-	public float MASS = 1;
+	public double MASS = 1;
 	
 	/**
 	 * Returns the mass of the galaxy. {@link Core#MASS} by default.
 	 * @return mass of the galaxy
 	 */
-	float getMass();
+	double getMass();
 
 }
