+++ src/ee/liiser/siim/galaxies/data/basic/BasicCore.java
@@ -1,6 +1,6 @@
 package ee.liiser.siim.galaxies.data.basic;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 import ee.liiser.siim.galaxies.calculations.Calculator;
 import ee.liiser.siim.galaxies.data.Core;
@@ -11,23 +11,23 @@ import ee.liiser.siim.galaxies.data.Core;
  */
 public class BasicCore extends BasicObject implements Core {
 	public BasicCore() {
-		this(new Vector3f(), new Vector3f());
+		this(new Vector3d(), new Vector3d());
 	}
 
-	public BasicCore(Vector3f position, Vector3f velocity) {
+	public BasicCore(Vector3d position, Vector3d velocity) {
 		this.position = position;
-		this.oldPosition = new Vector3f(velocity);
+		this.oldPosition = new Vector3d(velocity);
 		this.oldPosition.scale(-Calculator.dt);
 		this.oldPosition.add(position);
 	}
 
 	@Override
-	public float getSize() {
+	public double getSize() {
 		return CORE_SIZE;
 	}
 
 	@Override
-	public float getMass() {
+	public double getMass() {
 		return MASS;
 	}
 }
