+++ src/ee/liiser/siim/galaxies/data/velocity/VelocityObject.java
@@ -1,6 +1,6 @@
 package ee.liiser.siim.galaxies.data.velocity;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 import ee.liiser.siim.galaxies.data.BaseObject;
 
@@ -9,15 +9,15 @@ import ee.liiser.siim.galaxies.data.BaseObject;
  *
  */
 public abstract class VelocityObject extends BaseObject {
-	
-	protected Vector3f velocity;
-	
-	public Vector3f getVelocity(){
+
+	protected Vector3d velocity;
+
+	public Vector3d getVelocity() {
 		return velocity;
 	}
-	
-	public void setVelocity(Vector3f velocity){
+
+	public void setVelocity(Vector3d velocity) {
 		this.velocity = velocity;
 	}
-	
+
 }
